Laid down the Foundation for Armor and Relics
Created a functional Body and Head Slot function to only allow Shovel
Knight Armor and Relics to be Equipped.

Tested vs in game Armor and Head Items, and compatiablity with Mods that
increase Equipment slots (Backpacks and Amulets).  Everything checked
out!!!

Also readded the hold onto item if inventory is full function, makes
this Foundation feel correct with the game.